Restart docker daemon, when config updated.

+++ deployment/k8sPaiLibrary/maintaintool/docker-ce-install.sh
@@ -77,11 +77,14 @@ fi
     mkdir -p /etc/docker
 }
 
-[[ ! -f "/etc/docker/daemon.json" ]] &&
-{
+if [ ! -f "/etc/docker/daemon.json" ]; then
     cp $scriptPath/docker-daemon.json /etc/docker/daemon.json
-}
+    systemctl restart docker
+else
+    if python $scriptPath/docker-config-update.py -s $scriptPath/docker-daemon.json -d /etc/docker/daemon.json ; then
+        systemctl restart docker
+    else
+        echo There is no docker daemon configuration updated. Skip the process of restart docker daemon.
+    fi
+fi
 
-python $scriptPath/docker-config-update.py -s $scriptPath/docker-daemon.json -d /etc/docker/daemon.json
-
-systemctl restart docker
